add realistic boolean param to render_stitch_plan()
